kernel: Import the apple-gmux driver from Linux 4.8
Some Apple laptop models with dual GPUs use a special gmux device
in order to connect video outputs to a particular GPU or the other
one.
This driver allows the use of Intel graphics on Macbooks having an
on-board additional Nvidia GPU.

Rename the driver to "apple_gmux" since hyphens are not welcome in
DragonFly's kernel module names.