sbin/hammer: Make volume count test optional for blkdevs based commands
This is basically for feature enhancement of hammer strip command
in the next commit. No functional change by this commit.

Some blkdevs based commands need to be able to handle broken volume
headers by omitting volume verification (to certain extent). It has
been that way for the volume header contents, but not for volume
count as well as checking existence of root volume.