sbin/hammer: Allow any volume to be stripped by strip command
hammer strip should be able to strip each volume independently whether
all volumes that consist filesystem are available or not. It makes no
sense to require all volumes when the purpose of command is to destroy
them, even though that's necessary to go through the whole process.

If root volume is missing or root volume has incomplete offset format,
layer1/2 won't be erased, but just header part.