zoneinfo: Create a /usr/share/zoneinfo/UTC link.
We renamed GMT to UTC in 949d11f282cb90e67a6d24e8017b0ceb83746d29.
This commit is a followup to that commit and fixes tzload() when
called with gmt[] from gmtload().

Because upstream still uses GMT, stuff the link into a separate
file to not pollute the "etcetera" file.

Also, keep the /usr/share/zoneinfo/GMT link for now.

Reported-by: ftigeot