kernel/apple_gmux: Fix compilation when VGA_SWITCHEROO is in the config.
We would get a redefinition warning/error, because opt_drm.h has
VGA_SWITCHEROO defined to 1 in this case, which is not the same as
merely defining it. To be on the safe side, define to 1 here, too,
and also add an #ifndef and a comment.

Discussed-with: Peeter Must <karu.pruun@gmail.com>