sys/linker_set.h: Allow to declare empty data sets.
At first I thought it would be useful for OpenPAM static modules, but in case
when none of the static pam modules are linked in it is best to disable the
libpam OPENPAM_STATIC_MODULES option when it soon get enabled.

This weak version of SET_DECLARE() might still come in handy later on.
If set is empty, weak symbols of set begin/end pointers evaluate as 'NULL'
when they are not present. Also, this way we would make sure that in places
like bootloaders and kernel we are going to use non empty sets.