ncurses: Privatize the last public header.
Just move it with rest of privatized ncurses headers. Only user in base is
usr.sbin/installer/dfuife_curses/ and include path is already handled there.
As of this moment we are sure that none of private ncurses bits are used for
bootstrapping the world (non-interactive make/shell scripts redirected to
file should not care about colors anyway or try to format the layout).

No functional change.