bsd.sys.mk: Implement WARNS_AUDIT option.
The buildworld is expected not to pass this mode, so make(1) should
be invoked with '-k'. Strictly just as convenience for developers.
This is similar to FORMAT_AUDIT, for extra diagnostics by compiler at a
given WARNS level. The -Wall and -Wextra are getting bit too crowded.

While there, document FORMAT_AUDIT too.

In-discussion-with: swildner