newsyslog - Fix interval-based rotations when the -t flag is used
In this case, find the most-recently archived logfile
and use its mtime to determine whether or not to rotate,
as in the non-timestamped case.

Previously we would just try to use the mtime of <logfile>.0,
which always results in a rotation since it generally doesn't
exist in the -t case.

Obtained-from: FreeBSD (r248776)