loader/menu: Fix 'vfs.root.real_root' to be 'vfs.root.realroot'
This should be the same variable as used in /boot/loader.conf to specify
the real root partition.  See loader(8) and mkinitrd(8) for more info.

The intent here by unsetting it for the *rescue boot mode* is to prevent
the real root been mounted, and hence the system will stay in the md0
rescue root.

Verified-by: Matt Dillon