gcc80: Remove throw() from protos in libstdc++ headers.
The throw() as dynamic exception specificators was deprecated in c++11
and removed in c++17 (empty was made as an alias to noexcept(true)).

Our libc headers does not have the __attribute__(__nothrow__) markups.
To make libstdc++ fully usable for clangs on DragonFly it makes sense
to just remove the throw() additions in c++ language binding headers.
Previously in gcc50 we were excluding them just for __clang__ case as in:
838772be3c028f0cfd9fae5da181858e4199863a