kernel/nfs: Use the correct prototype for sys_nfssvc().
The code handles loading of nfs.ko (nfssvc() is the only {,un}loadable
syscall). sys_nfssvc()'s prototype is automatically generated, so just
use that.

The sy_narg setting is not strictly needed because it is already
initialized correctly. Just the sy_call pointer is either sys_nosys
or sys_nfssvc, depending on the module load.

Reported-by: zrj