btxld(8): Convert to hostprog.
There is no point to provide this utility in base. We are not planing to
bring back i386 support. Move it into sys/boot/pc32/ for easier handling.

This way the AOUT_H_FORCE32 now is confined to the same pc32/ directory.
The <sys/imgact_aout.h> has override for it to create proper btx header.

For future we should make btxld endian neutral.