Makefile.inc1: Do not continue the build upon an error in various cases.
Before this commit, the build would limp on even though a library
failed building, for example, leading to sometimes confusing followup
errors later in the build. It was also more difficult to see directly
why a build really failed in the first place.

To prevent this, make various targets use '&&' instead of ';' to
chain commands.

People wanting the build to carry on for as long as possible, can
always use make's -k option.