In several standard headers, harmonize checks for legacy defines etc.
When something goes away in a certain issue of the standard, always
check with < $standard, not <= $previous_standard. Also, uniformly
check for __XSI_VISIBLE or __POSIX_VISIBLE being > 0 in these cases.

While here, fix a small nit in <sys/wait.h>: wait3() was removed
from Issue 6, but the requirement to define the rusage structure was
removed from Issue 7.