kernel: Remove unused *.h files from SRCS in kernel module Makefiles.
They were found by checking the preprocessed code of the files
in SRCS to see if the header was included at some point.

After removal, the preprocessed source of a build with the old
Makefiles was compared against one with the changes (for various
kernel configurations and when building just from /usr/src) to
verify that the commit leads to no functional change.