nrelease: Use part-by-label to identify root in .img
Previously, the DragonFly .img hardcoded "da8" as the USB device, but
this may not be the case on some systems and requires user interaction
to boot.

Now that DragonFly has "/dev/part-by-label", use this new mechanism
(i.e., the label/name of the disklabel64(5) slice in the .img file) to
better identify the USB device with DragonFly installation image on it.
The label name is "DragonFly_${KERNCONF}" by default.