ministat(1): Fix terminal width handling
ministat(1) will prefer to get the terminal width from the "COLUMNS"
environment variable, in which case the plot will overflow by 2 columns
because the plot frame is not considered.  Fix this issue.

However, in Emacs's "shell" mode (not the "term" mode), the plot will
still overflow because Emacs puts a '\' at the right-most column.  One
can manually modify the "COLUMNS" environment variable or use the "-w"
option to workaround this issue.

Reported-by: swildner